# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Genrel Api",
        "description": "Apis for Goodreads Website!",
        "version": "1.0.0"
    },
    "servers": [{
        "url": "http://localhost:3000/"
    }],
    "security": [{
        "bearerAuth": []
    }],
    "paths": {
...
        "/genre": {
            "get": {
                "tags": [
                    "Genre"
                ],
                "summary": "Get All Genres",
                "responses": {
                    "200": {
                        "description": "all users"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/book/add-book": {
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Add Books.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "Gener_id": {
                                        "type": "number"
                                    },
                                    "ISBN": {
                                        "type": "string"
                                    },
                                    "AvgRating": {
                                        "type": "number"
                                    },
                                    "Title": {
                                        "type": "string"
                                    },
                                    "Author": {
                                        "type": "string"
                                    },
                                    "Description": {
                                        "type": "string"
                                    },
                                    "Publication-year": {
                                        "type": "integer"
                                    },
                                    "image-url": {
                                        "type": "string"
                                    },
                                    "image-url-s": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book Added"
                    },
                    "401": {
                        "description": "Auth Fail"
                    },
                    "409": {
                        "description": "Book All Ready Exist"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/book/edit-book": {
            "patch": {
                "tags": [
                    "Book"
                ],
                "summary": "Edit Books.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "Gener_id": {
                                        "type": "number"
                                    },
                                    "ISBN": {
                                        "type": "string"
                                    },
                                    "AvgRating": {
                                        "type": "number"
                                    },
                                    "Title": {
                                        "type": "string"
                                    },
                                    "Author": {
                                        "type": "string"
                                    },
                                    "Description": {
                                        "type": "string"
                                    },
                                    "Publication-year": {
                                        "type": "integer"
                                    },
                                    "image-url": {
                                        "type": "string"
                                    },
                                    "image-url-s": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book Updated"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/book/delete/isbn/{isbn}": {
            "delete": {
                "tags": [
                    "Book"
                ],
                "summary": "Delete a Book.",
                "parameters": [{
                    "name": "isbn",
                    "in": "path",
                    "description": "Book ISBN",
                    "required": true,
                    "style": "simple",
                    "explode": false,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Book deleted"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/book/delete/id/{id}": {
            "delete": {
                "tags": [
                    "Book"
                ],
                "summary": "Delete a Book.",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "Book ID",
                    "required": true,
                    "style": "simple",
                    "explode": false,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Book deleted"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/book": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get all Books. Limit to 10",
                "responses": {
                    "200": {
                        "description": "Get Books"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/book/id/{id}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get a Book by id",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "Book Id",
                    "required": true,
                    "style": "simple",
                    "explode": false,
                    "schema": {
                        "type": "number"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Get Books"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/book/ISBN/{isb}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Get a Book by isbn",
                "parameters": [{
                    "name": "isb",
                    "in": "path",
                    "description": "Book ISBN",
                    "required": true,
                    "style": "simple",
                    "explode": false,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Get Books"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/book/author/": {
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Get a Book by author",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Get Books"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/book/title": {
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Get a Book by title",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Get Books"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/book/genre": {
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Get a Book by genre",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "Genre_id": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Get Books"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/book/search": {
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Search Book by ISBN,Title,Author",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "s": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Get Books"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}